스터디 화면 조회용 dto를 하나의 dto로 합쳐라
화면을 조회하기 위한 스터디 dto를 하나로 합칩니다
