+++ app/src/main/java/com/example/bookclub/application/StudyCommentService.java
@@ -3,7 +3,7 @@ package com.example.bookclub.application;
 import com.example.bookclub.domain.Study;
 import com.example.bookclub.domain.StudyComment;
 import com.example.bookclub.domain.StudyCommentRepository;
-import com.example.bookclub.dto.StudyDto;
+import com.example.bookclub.dto.StudyApiDto;
 import com.example.bookclub.errors.StudyCommentContentNotExistedException;
 import com.example.bookclub.errors.StudyCommentDeleteBadRequest;
 import com.example.bookclub.errors.StudyCommentNotFoundException;
@@ -36,8 +36,8 @@ public class StudyCommentService {
 	 * @throws StudyCommentContentNotExistedException 생성할 댓글 정보가 비어있는 경우
 	 */
 	@Transactional
-	public StudyDto.StudyCommentResultDto createStudyComment(UserAccount userAccount, Long studyId,
-															 StudyDto.StudyCommentCreateDto studyCommentCreateDto) {
+	public StudyApiDto.StudyCommentResultDto createStudyComment(UserAccount userAccount, Long studyId,
+																StudyApiDto.StudyCommentCreateDto studyCommentCreateDto) {
 		Study study = studyService.getStudy(studyId);
 		if(studyCommentCreateDto.getContent().isBlank()) {
 			throw new StudyCommentContentNotExistedException();
@@ -47,7 +47,7 @@ public class StudyCommentService {
 
 		StudyComment savedStudyComment = studyCommentRepository.save(studyComment);
 
-		return StudyDto.StudyCommentResultDto.of(savedStudyComment, userAccount.getAccount());
+		return StudyApiDto.StudyCommentResultDto.of(savedStudyComment, userAccount.getAccount());
 	}
 
 	/**
@@ -59,7 +59,7 @@ public class StudyCommentService {
 	 * @throws StudyCommentDeleteBadRequest 스터디 댓글이 로그인한 사용자의 댓글이 아닌 경우
 	 */
 	@Transactional
-	public StudyDto.StudyCommentResultDto deleteStudyComment(UserAccount userAccount, long studyCommentId) {
+	public StudyApiDto.StudyCommentResultDto deleteStudyComment(UserAccount userAccount, long studyCommentId) {
 		StudyComment studyComment = getStudyComment(studyCommentId);
 		if(!studyComment.getAccount().getId().equals(userAccount.getAccount().getId())) {
 			throw new StudyCommentDeleteBadRequest();
@@ -67,7 +67,7 @@ public class StudyCommentService {
 
 		studyCommentRepository.delete(studyComment);
 
-		return StudyDto.StudyCommentResultDto.of(studyComment, userAccount.getAccount());
+		return StudyApiDto.StudyCommentResultDto.of(studyComment, userAccount.getAccount());
 	}
 
 	/**
