+++ app/src/main/java/com/example/bookclub/application/StudyService.java
@@ -4,8 +4,8 @@ import com.example.bookclub.domain.Account;
 import com.example.bookclub.domain.Study;
 import com.example.bookclub.domain.StudyComment;
 import com.example.bookclub.domain.StudyState;
+import com.example.bookclub.dto.StudyApiDto;
 import com.example.bookclub.dto.StudyDto;
-import com.example.bookclub.dto.StudyInfoResultDto;
 import com.example.bookclub.errors.AccountNotManagerOfStudyException;
 import com.example.bookclub.errors.ParseTimeException;
 import com.example.bookclub.errors.StudyAlreadyExistedException;
@@ -60,7 +60,7 @@ public class StudyService {
      * @throws StudyStartAndEndDateNotValidException 생성하려는 스터디 종료일이 시작일보다 빠른 경우
      * @throws StudyStartAndEndTimeNotValidException 생성하려는 스터디 종료시간이 시작시간보다 빠른 경우
      */
-    public StudyDto.StudyResultDto createStudy(String email, StudyDto.StudyCreateDto studyCreateDto) {
+    public StudyApiDto.StudyResultDto createStudy(String email, StudyApiDto.StudyCreateDto studyCreateDto) {
         Account loginAccount = accountService.findAccountByEmail(email);
 
         if(loginAccount.getStudy() != null) {
@@ -90,7 +90,7 @@ public class StudyService {
         study.addAdmin(loginAccount);
         Study createdStudy = studyRepository.save(study);
 
-        return StudyDto.StudyResultDto.of(createdStudy);
+        return StudyApiDto.StudyResultDto.of(createdStudy);
     }
 
     /**
@@ -105,7 +105,7 @@ public class StudyService {
      * @throws StudyStartAndEndDateNotValidException 스터디 식별자에 해당하는 스터디 종료일이 시작일보다 빠른 경우
      * @throws StudyStartAndEndTimeNotValidException 스터디 식별자에 해당하는 스터디 종료시간이 시작시간보다 빠른 경우
      */
-    public StudyDto.StudyResultDto updateStudy(String email, Long id, StudyDto.StudyUpdateDto studyUpdateDto) {
+    public StudyApiDto.StudyResultDto updateStudy(String email, Long id, StudyApiDto.StudyUpdateDto studyUpdateDto) {
         Study study = getStudy(id);
         Account loginAccount = accountService.findAccountByEmail(email);
         if(!study.getEmail().equals(loginAccount.getEmail())) {
@@ -127,7 +127,7 @@ public class StudyService {
         }
 
         study.updateWith(studyUpdateDto);
-        return StudyDto.StudyResultDto.of(study);
+        return StudyApiDto.StudyResultDto.of(study);
     }
 
     /**
@@ -179,7 +179,7 @@ public class StudyService {
      * @return 삭제된 스터디 식별자
      * @throws AccountNotManagerOfStudyException 스터디 식별자에 해당하는 스터디 이메일과 사용자 식별자가 다른 경우
      */
-    public StudyDto.StudyResultDto deleteStudy(String email, Long id) {
+    public StudyApiDto.StudyResultDto deleteStudy(String email, Long id) {
         Study study = getStudy(id);
         Account loginAccount = accountService.findAccountByEmail(email);
         if(!study.getEmail().equals(loginAccount.getEmail())) {
@@ -189,7 +189,7 @@ public class StudyService {
         study.deleteAccounts();
         studyRepository.delete(study);
 
-        return StudyDto.StudyResultDto.of(study);
+        return StudyApiDto.StudyResultDto.of(study);
     }
 
     /**
@@ -202,7 +202,7 @@ public class StudyService {
      * @throws StudyAlreadyExistedException 스터디 식별자에 해당하는 스터디 지원이 이미 존재하는 경우
      * @throws StudySizeFullException 스터디 식별자에 해당하는 스터디 정원이 다 찬 경우
      */
-    public StudyDto.StudyApplyResultDto applyStudy(UserAccount userAccount, Long id) {
+    public StudyApiDto.StudyApplyResultDto applyStudy(UserAccount userAccount, Long id) {
         Study study = getStudy(id);
         Account account = userAccount.getAccount();
 
@@ -220,7 +220,7 @@ public class StudyService {
 
         study.addAccount(account);
 
-        return StudyDto.StudyApplyResultDto.of(id);
+        return StudyApiDto.StudyApplyResultDto.of(id);
     }
 
     /**
@@ -232,7 +232,7 @@ public class StudyService {
      * @throws StudyNotInOpenStateException 스터디 식별자에 해당하는 스터디가 모집중이 아닌 경우
      * @throws StudyNotAppliedBefore 스터디 식별자에 해당하는 스터디 신청이 존재하지 않는 경우
      */
-    public StudyDto.StudyApplyResultDto cancelStudy(UserAccount userAccount, Long id) {
+    public StudyApiDto.StudyApplyResultDto cancelStudy(UserAccount userAccount, Long id) {
         Study study = getStudy(id);
         Account account = userAccount.getAccount();
 
@@ -246,7 +246,7 @@ public class StudyService {
 
         study.cancelAccount(account);
 
-        return StudyDto.StudyApplyResultDto.of(id);
+        return StudyApiDto.StudyApplyResultDto.of(id);
     }
 
     /**
@@ -278,7 +278,7 @@ public class StudyService {
      * @param id 스터디 식별자
      * @return 스터디 식별자에 해당하는 스터디 정보
      */
-    public StudyDto.StudyDetailResultDto getDetailedStudy(UserAccount userAccount, Long id) {
+    public StudyApiDto.StudyDetailResultDto getDetailedStudy(UserAccount userAccount, Long id) {
         Long principalId = userAccount.getAccount().getId();
         Study study = getStudy(id);
         List<StudyComment> studyComments = study.getStudyComments();
@@ -300,13 +300,13 @@ public class StudyService {
                 studyComment.setIsWrittenByMeTrue();
         });
 
-        List<StudyDto.StudyCommentResultDto> studyCommentResultDtos = studyComments.stream()
+        List<StudyApiDto.StudyCommentResultDto> studyCommentResultDtos = studyComments.stream()
                 .map(studyComment -> {
-                        return StudyDto.StudyCommentResultDto.of(studyComment, studyComment.getAccount());
+                        return StudyApiDto.StudyCommentResultDto.of(studyComment, studyComment.getAccount());
                     })
                 .collect(Collectors.toList());
 
-        return StudyDto.StudyDetailResultDto.of(StudyDto.StudyResultDto.of(study), studyCommentResultDtos);
+        return StudyApiDto.StudyDetailResultDto.of(StudyApiDto.StudyResultDto.of(study), studyCommentResultDtos);
     }
 
     /**
@@ -318,7 +318,7 @@ public class StudyService {
      * @param pageable 페이징 정보
      * @return 검색어와 스터디 상태에 해당하는 스터디 페이징 정보
      */
-    public Page<StudyDto.StudyResultDto> getStudiesBySearch(String keyword, StudyState studyState, Long principalId, Pageable pageable) {
+    public Page<StudyApiDto.StudyResultDto> getStudiesBySearch(String keyword, StudyState studyState, Long principalId, Pageable pageable) {
         List<Study> studies = studyRepository.findByBookNameContaining(keyword, studyState, pageable);
         long total = studyRepository.getStudiesCountByKeyword(keyword, studyState);
 
@@ -331,8 +331,8 @@ public class StudyService {
             });
         });
 
-        List<StudyDto.StudyResultDto> studyResultDtos = studies.stream()
-                .map(StudyDto.StudyResultDto::of)
+        List<StudyApiDto.StudyResultDto> studyResultDtos = studies.stream()
+                .map(StudyApiDto.StudyResultDto::of)
                 .collect(Collectors.toList());
 
         return new PageImpl<>(studyResultDtos, pageable, total);
@@ -345,7 +345,7 @@ public class StudyService {
      * @param id 스터디 식별자
      * @return 스터디 식별자에 해당하는 스터디 정보
      */
-    public StudyInfoResultDto getStudyInfo(Long id) {
+    public StudyDto.StudyInfoResultDto getStudyInfo(Long id) {
         return studyRepository.getStudyInfo(id);
     }
