--- app/src/main/java/com/example/bookclub/dto/StudyAccountInfoResultDto.java
@@ -1,20 +1,0 @@
-package com.example.bookclub.dto;
-
-import com.querydsl.core.annotations.QueryProjection;
-import lombok.Getter;
-
-@Getter
-public class StudyAccountInfoResultDto {
-	private String name;
-
-	private String email;
-
-	private String nickname;
-
-	@QueryProjection
-	public StudyAccountInfoResultDto(String name, String email, String nickname) {
-		this.name = name;
-		this.email = email;
-		this.nickname = nickname;
-	}
-}
